split _shared
