+++ .github/workflows/check-if-action-bindings-up-to-date.main.kts
@@ -1,7 +1,7 @@
 #!/usr/bin/env kotlin
 @file:Repository("https://s01.oss.sonatype.org/content/repositories/snapshots/")
 @file:DependsOn("io.github.typesafegithub:github-workflows-kt:1.4.1-20231029.193015-11")
-@file:Import("_shared.main.kts")
+@file:Import("setup-java.main.kts")
 @file:Import("generated/actions/checkout.kt")
 @file:Import("generated/gradle/gradle-build-action.kt")
